requirements.txt
torch
transformers
spacy 
scispacy 
en_core_sci_sm
flask
requests
beautifulsoup4
nltk
pandas
scikit-learn
tqdm
python-dotenv
flask-cors
pytest
validators
pillow
accelerate
PyPDF2>=3.0.0
pytesseract>=0.3.10
tabula-py>=2.7.0
pdf2image>=1.16.3
# Hardware-specific extensions will be installed automatically by run.bat
# - Intel CPUs: intel-extension-for-pytorch
# - AMD GPUs: ROCm-enabled PyTorch (if available)